Allow methods to be called without triggering event callbacks
This adds a `skipCb` parameter to open(), save(), close(), and reset().
If `skipCB` is true, the event callback for that method will not be
called.

The main purpose of this is to allow save() to be called from within the
save event callback, without causing a recursive call to the save event
callback. The same functionality is added to the other methods for
completeness.